+++ src/index.js
@@ -108,10 +108,7 @@ export function concatClassNames(value) {
 // "msTransition" => "-ms-transition"
 function hyphenateStyleName(styleName) {
   if (!styleNameCache.has(styleName)) {
-    const name = styleName
-      .replace(uppercasePattern, '-$&')
-      .toLowerCase()
-      .replace(msPattern, '-ms-')
+    const name = styleName.replace(uppercasePattern, '-$&').toLowerCase().replace(msPattern, '-ms-')
 
     // returns 'undefined' instead of the 'Map' object in IE11
     styleNameCache.set(styleName, name)
@@ -253,11 +250,13 @@ export function renderer(view, state, actions) {
               children: node,
               footer: '',
             })
+          } else if (node.tag === 3) {
+            out += escapeHtml(node.type)
           } else if (node.type === 3) {
             out += escapeHtml(node.name)
           } else if (typeof node === 'object') {
             out += renderFragment(
-              node.name || node.nodeName,
+              node.name || node.type || node.nodeName,
               node.props || node.attributes,
               node.children,
               stack,