+++ test/index.test.js
@@ -449,7 +449,7 @@ describe('renderToString(view, state, actions)', () => {
     expect(html).toBe('<meta/><link/>')
   })
 
-  it('should support Hyperapp V2', () => {
+  it('should support Hyperapp v2.0.0', () => {
     const VNode = {
       name: 'div',
       props: {},
@@ -466,6 +466,28 @@ describe('renderToString(view, state, actions)', () => {
       element: null,
       key: null,
       type: 1,
+    }
+    const html = renderToString(VNode)
+    expect(html).toBe('<div>foo bar baz</div>')
+  })
+
+  it('should support Hyperapp v2.0.6', () => {
+    const VNode = {
+      type: 'div',
+      props: {},
+      children: [
+        {
+          type: 'foo bar baz',
+          props: {},
+          children: [],
+          node: null,
+          key: null,
+          tag: 3,
+        },
+      ],
+      node: null,
+      key: null,
+      tag: 1,
     }
     const html = renderToString(VNode)
     expect(html).toBe('<div>foo bar baz</div>')